+++ NEWS.md
@@ -1,3 +1,5 @@
+# ospsuite (development version)
+
 # ospsuite 12.4.4
 
 ## Minor improvements and bug fixes